Refactor the test helpers; Extract michelson maps helpers to a separate module;
