+++ tests/helpers/index.ts
@@ -5,3 +5,4 @@ export * from './useLastTezosToolkit';
 export * from './contractDeployment';
 export * from './utils';
 export * from './signing';
+export * from './michelsonMaps';
